Weather_Bot_using_GoogleDialogFlow/requirements.txt
blinker==1.9.0
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Cors==5.0.0
geojson==2.5.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
pyowm==3.3.0
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.8.0
six==1.17.0
urllib3==2.3.0
Werkzeug==3.1.3
wincertstore==0.2
-e .